requirements.txt
tensorflow==1.14.0
keras==2.0.8
protobuf==3.20.1
gradio==3.0.15
gdown==4.4.0
numpy
scipy
Pillow
cython
matplotlib
scikit-image
opencv-python
h5py==2.10.0
imgaug
IPython[all]
